% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert_cameo.R
\name{convert_cameo}
\alias{convert_cameo}
\alias{countrycode}
\title{Convert CAMEO Codes}
\usage{
convert_cameo(cameo)
}
\arguments{
\item{cameo}{Vector of CAMEO event codes.}
}
\description{
Converts CAMEO codes. CAMEO is an event ontology used in event data projects, including Phoenix.
}
\examples{
events$Description <- phoxy::convert_cameo(events$EventCode) # Vector of values to be converted
}
\keyword{data}
\keyword{event}

