% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert_cameo_data.R
\docType{data}
\name{convert_cameo_data}
\alias{convert_cameo_data}
\title{CAMEO code translation data frame}
\format{A data frame with 310 rows and 2 columns}
\usage{
convert_cameo_data
}
\description{
A data frame with 310 rows and 2 columns.
Used internally by the \code{convert_cameo()} function.
}
\details{
\itemize{
  \item CAMEOcode: the 310 different low-level CAMEO codes.
  \item EventDescription: Human-readable descriptions of the codes.
}
}
\note{
The current CAMEO codebook is located here: \url{http://eventdata.parusanalytics.com/data.dir/cameo.html}.
}
\keyword{datasets}

