% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_icews.R
\name{download_icews}
\alias{download_icews}
\title{Download the ICEWS Dataset}
\usage{
download_icews(destpath = ".")
}
\arguments{
\item{destpath}{The path to the directory where ICEWS should go.}
}
\description{
Download and unzip all of the data files for the ICEWS dataset from the
Harvard Dataverse into a given directory.
}
\note{
This function is still in development and may contain errors and change quickly.
}
\examples{
download_icews("~/ICEWS/")
}
\author{
Tony Boyles
}

