% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_links}
\alias{get_links}
\title{Download the Phoenix Dataset}
\usage{
get_links()
}
\arguments{
\item{dir}{The path to download Phoenix into.}

\item{version}{[Not yet implemented]. Download a specific version of Phoenix.}
}
\description{
Download and unzip all of the data files for the Phoenix dataset from the
Phoenix data website into a given directory.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{
download_phoenix("~/OEDA/phoxy_test/")
}
\author{
Andy Halterman
}

