% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ingest_icews.R
\name{ingest_icews}
\alias{ingest_icews}
\title{Ingest the ICEWS Event Dataset}
\usage{
ingest_icews(dir, read_func = "read.csv", processing_function)
}
\arguments{
\item{dir}{The path to the ICEWS folder.}

\item{read_func}{[Not yet implemented]. Use an alternative reading function like \code{data.table::fread} or \code{readr::read_delim}.}
}
\value{
A single dataframe with all the ICEWS events in the folder.
}
\description{
Given a directory with individual ICEWS dataset files, quickly read
them all in, name them correctly, and combine them into one dataframe.
}
\note{
This function is still in development and may contain errors and change quickly.
}
\examples{
events <- ingest_icews("~/ICEWS/study_28075/Data/")
}
\author{
Andy Halterman
}

