% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{phoxy}
\alias{phoxy}
\alias{phoxy-package}
\title{phoxy: download, ingest, and analyze the Phoenix event dataset}
\description{
The Phoenix event data set records social and political events
automatically extracted from news reporting. \code{phoxy} provides
functions to make acquiring and analyzing this event data easier. For more
information about Phoenix, see its website here:
  \url{http://phoenixdata.org/} and the website of the Open Event Data
Alliance here: \url{http://openeventdata.org/}.
}

