% Generated by roxygen2 (4.0.1): do not edit by hand
\name{update_phoenix}
\alias{update_phoenix}
\title{Update a local directory of Phoenix dataset files with new files from the server}
\usage{
update_phoenix()
}
\description{
Checks the contents of a directory containing Phoenix event data files, checks whether the
server has new events, and downloads them to that directory. (It'll have some version handling ability,
too, either from the file names or by reading in the events.)
}
\details{
#'
}

