% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download_phoenix.R
\name{get_links}
\alias{get_links}
\title{Download the Phoenix Dataset}
\usage{
get_links(phoenix_version)
}
\arguments{
\item{destpath}{The path to the directory where Phoenix should go.}

\item{phoenix_version.}{Download a specific version of Phoenix ("v0.1.0" or "current").}
}
\description{
Download and unzip all of the data files for the Phoenix dataset from the
Phoenix data website into a given directory.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{
download_phoenix("~/OEDA/phoxy_test/", phoenix_version = "current")
}
\author{
Andy Halterman
}

