% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_phoenix.R
\name{download_phoenix}
\alias{download_phoenix}
\title{Download the Phoenix Dataset}
\usage{
download_phoenix(destpath, phoenix_version = "current", start_date = NULL,
  end_date = NULL)
}
\arguments{
\item{destpath}{The path to the directory where Phoenix should go.}

\item{phoenix_version}{Download a specific version of Phoenix ("v0.1.0" or the current version by default).}

\item{start_date}{Filter the dataset to only include events from start_date.}

\item{end_date}{Filter the dataset to only include events before end_date.}
}
\description{
Download and unzip all of the data files for the Phoenix dataset from the
Phoenix data website into a given directory.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{

download_phoenix("~/OEDA/phoxy_test/", phoenix_version = "current", start_date = "2017-01-01", end_date = "2017-01-31")

}
\author{
Andy Halterman, Altaf Ali
}
