% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ingest_phoenix}
\alias{ingest_phoenix}
\title{Ingest the Phoenix Dataset}
\usage{
ingest_phoenix(dir, version = "auto", read_func = "read.csv")
}
\arguments{
\item{dir}{The path to the Phoenix folder.}

\item{version}{[Not yet implemented]. Use the appropriate function for each Phoenix version.}

\item{read_func}{[Not yet implemented]. Use an alternative reading function like \code{fread} or \code{read_csv}.}
}
\value{
A single dataframe with all the Phoenix events in the folder.
}
\description{
Given a directory with individual Phoenix dataset files, quickly read
them all in, name them correctly, and combine them into one dataframe.
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{
events <- ingest_phoenix("~/OEDA/phoxy_test/")
}
\author{
Andy Halterman
}

