% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/update_phoenix.R
\name{update_phoenix}
\alias{update_phoenix}
\title{Update a local directory of Phoenix dataset files with new files from the server}
\usage{
update_phoenix(destpath, phoenix_version = "current")
}
\arguments{
\item{destpath}{The path to download Phoenix into.}

\item{phoenix_version.}{Download a specific version of Phoenix ("v0.1.0" or "current").}
}
\description{
Checks the contents of a directory containing Phoenix event data files, checks whether the
server has new events, and downloads them to that directory. (It'll have some version handling ability,
too, either from the file names or by reading in the events.)
}
\note{
This function, like Phoenix, is still in development and may contain errors and change quickly.
}
\examples{
update_phoenix("~/OEDA/phoxy_test/", phoenix_version = "current")
}
\author{
Andy Halterman
}

